tests viewer okay
