+++ src/opengeodeweb_viewer/object/methods.py
@@ -1,5 +1,5 @@
 # Standard library imports
-
+import os
 
 # Third party imports
