+++ src/opengeodeweb_viewer/rpc/mesh/protocols.py
@@ -11,7 +11,7 @@ from wslink import register as exportRpc
 
 # Local application imports
 from opengeodeweb_viewer.utils_functions import get_schemas_dict, validate_schema
-from opengeodeweb_viewer.vtk_protocol import VtkView
+from opengeodeweb_viewer.object.methods import VtkObjectView
 
 
 schemas_dir = os.path.join(os.path.dirname(__file__), "schemas")
@@ -23,35 +23,38 @@ class VtkMeshView(VtkObjectView):
 
     @exportRpc(schemas_dict["toggle_object_visibility"]["rpc"])
     def toggle_object_visibility(self, params):
-        print(f"{schemas_dict["toggle_object_visibility"]["rpc"]=}", flush=True)
+        print(schemas_dict["toggle_object_visibility"]["rpc"], flush=True)
         validate_schema(params, schemas_dict["toggle_object_visibility"])
         super().SetVisibility(params)
 
     @exportRpc(schemas_dict["set_opacity"]["rpc"])
     def SetOpacity(self, params):
-        print(f"{schemas_dict["set_opacity"]["rpc"]=}", flush=True)
+        print(schemas_dict["set_opacity"]["rpc"], flush=True)
         validate_schema(params, schemas_dict["set_opacity"])
         super().SetOpacity(params)
 
     @exportRpc(schemas_dict["toggle_edge_visibility"]["rpc"])
     def setEdgeVisibility(self, params):
-        print(f"{schemas_dict["toggle_edge_visibility"]["rpc"]=}", flush=True)
+        print(schemas_dict["toggle_edge_visibility"]["rpc"], flush=True)
         validate_schema(params, schemas_dict["toggle_edge_visibility"])
         print(f"{params=}", flush=True)
         super().SetEdgeVisibility(params)
 
     @exportRpc(schemas_dict["toggle_point_visibility"]["rpc"])
     def setPointVisibility(self, params):
+        print(schemas_dict["toggle_point_visibility"]["rpc"], flush=True)
         validate_schema(params, schemas_dict["toggle_point_visibility"])
         super().SetVertexVisibility(params)
 
     @exportRpc(schemas_dict["set_point_size"]["rpc"])
     def setPointSize(self, params):
+        print(schemas_dict["set_point_size"]["rpc"], flush=True)
         validate_schema(params, schemas_dict["set_point_size"])
         super().SetPointSize(params)
 
     @exportRpc(schemas_dict["set_color"]["rpc"])
     def setColor(self, params):
+        print(schemas_dict["set_color"]["rpc"], flush=True)
         validate_schema(params, schemas_dict["set_color"])
         super().SetColor(params)
 
@@ -69,7 +72,7 @@ class VtkMeshView(VtkObjectView):
 
     @exportRpc(schemas_dict["display_polygon_attribute"]["rpc"])
     def setPolygonAttribute(self, id, name):
-        store = self.getObject(id)
+        validate_schema(params, schemas_dict["display_polygon_attribute"])
         cpp = store["cpp"]
         print(cpp.nb_polygons())
         cells = store["vtk"].GetCellData()
@@ -81,7 +84,7 @@ class VtkMeshView(VtkObjectView):
             data = self.computeAttributeData(manager, name)
             cells.AddArray(data)
         cells.SetActiveScalars(name)
-        mapper = store["mapper"]
+        mapper = self.get_object(id)["mapper"]
         mapper.ScalarVisibilityOn()
         mapper.SetScalarModeToUseCellData()
         mapper.SetScalarRange(cells.GetScalars().GetRange())
