+++ src/tests/data/images/viewer/apply_textures.jpeg
