+++ src/tests/data/images/viewer/create_object_pipeline.jpeg
