+++ src/tests/data/images/viewer/create_visualization.jpeg
