+++ src/tests/data/images/viewer/delete_object_pipeline.jpeg
