+++ src/tests/data/images/viewer/reset.jpeg
