+++ src/tests/data/images/viewer/reset_camera.jpeg
