+++ src/tests/data/images/viewer/set_viewer_background_color.jpeg
