+++ src/tests/data/images/viewer/toggle_object_visibility_1.jpeg
