+++ src/tests/data/images/viewer/toggle_object_visibility_2.jpeg
