+++ src/tests/data/images/viewer/toggle_object_visiblity.jpeg
