+++ src/tests/data/take_screenshot_with_background.jpg
