+++ src/tests/data/take_screenshot_without_background.png
